+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # RoleModel/actions
 
+## [v3.2.0] Mar 5, 2026
+
+- Add new `staging-auto-merge` action, migrated from https://github.com/RoleModel/staging-auto-merge.
+  - Update staging conflict handling to remove the `Staging` label immediately when merge conflicts occur.
+  - Update staging squash commit messages to omit PR numbers.
+
 ## [v3.1.0] Feb 26, 2026
 
 Add new `project-stats` action which provides a markdown version of
@@ -39,6 +45,7 @@ New Composite Actions:
 - Add `test-cleanup` action to save logs
 - Add `test-runtime-analyzer` action to post test runtimes
 
+[v3.2.0]: https://github.com/RoleModel/actions/releases/tag/v3.2.0
 [v3.1.0]: https://github.com/RoleModel/actions/releases/tag/v3.1.0
 [v3.0.0]: https://github.com/RoleModel/actions/releases/tag/v3.0.0
 [v2.0.0]: https://github.com/RoleModel/actions/releases/tag/v2.0.0